requirements.txt
SpeechRecognition
python_video_converter
os
sys